% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{mountain_colours}
\alias{mountain_colours}
\title{Mountain Colours}
\format{
An object of class \code{list} of length 6.
}
\usage{
mountain_colours
}
\description{
Mountain Colours
}
\examples{
mountain_colours
}
\keyword{datasets}
