% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{mountain_palette}
\alias{mountain_palette}
\title{Mountain Palette}
\usage{
mountain_palette(primary = "navajowhite", other = "darkkhaki", direction = 1)
}
\arguments{
\item{primary}{The primary colour in the palette which is navajowhite}

\item{other}{In case there are only 2 colours to be chosen then, this colour will be used. It is darkkhaki}

\item{direction}{If value of direction is 0 or greater than 0, then the colours will be in the order: navajowhite, lightsteelblue, silver, lightpink, sandybrown, darkkhaki. If it is less than 1 then the direction will be reversed.}
}
\value{
The color list with n number of colours where n is the parameter passed
}
\description{
Mountain Palette
}
\examples{
mountain_palette()(3)
mountain_palette("lightsteelblue", "lightpink", -1)(2)
}
