% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_mountain_d}
\alias{scale_colour_mountain_d}
\title{Discrete Scale for color from the mountain palette}
\usage{
scale_colour_mountain_d(
  primary = "navajowhite",
  other = "darkkhaki",
  direction = 1,
  ...
)
}
\arguments{
\item{primary}{The primary colour in the palette which is navajowhite, other colours from the palette can also be given}

\item{other}{In case there are only 2 colours to be chosen then, this colour will be used. It is darkkhaki}

\item{direction}{If value of direction is 0 or greater than 0, then the colours will be in the order: navajowhite, lightsteelblue, silver, lightpink, sandybrown, darkkhaki. If it is less than 1 then the direction will be reversed.}

\item{...}{Other parameters that support the scale_fill_discrete function}
}
\value{
A ggplot2 object containing the scaled colours
}
\description{
Discrete Scale for color from the mountain palette
}
\examples{
scale_colour_mountain_d()
scale_colour_mountain_d(primary = "silver", direction = -1)
}
