% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_mountain_seq_c}
\alias{scale_colour_mountain_seq_c}
\title{Continuous Scale for color from the mountain palette}
\usage{
scale_colour_mountain_seq_c(
  name = "darkkhaki",
  direction = 1,
  amount = 0.5,
  ...
)
}
\arguments{
\item{name}{The name of the colour from the palette, the default colour is darkkhaki.}

\item{direction}{If the value is greater than or equal to 0 then lighter colours correspond to higher values. Otherwise lighter colours will correspond to lower values}

\item{amount}{The luminance of the colour chosen will change according to the value given. If value is negative then it will make the scale more darker.}

\item{...}{Other parameters that support the scale_fill_continuous function}
}
\value{
A ggplot2 object containing the scaled colours
}
\description{
Continuous Scale for color from the mountain palette
}
\examples{
scale_colour_mountain_seq_c()
scale_colour_mountain_seq_c(name = "silver", direction = -1, amount = -0.7)
}
